requirements.txt
transformers
langchain-community
huggingface_hub
TensorFlow
tf-keras
chromadb==0.4.6
InstructorEmbedding
sentence-transformers==2.2.2
faiss-cpu
docx2txt
unstructured
unstructured[pdf]
PyPDF2
langchain_huggingface 
pydantic
streamlit
streamlit_extras